examples/foundational/requirements.txt
fastapi
uvicorn
python-dotenv
pipecat-ai[webrtc,daily,deepgram,cartesia,openai,azure,silero]
pipecat-ai-small-webrtc-prebuilt
